+++ test/regress/test-s2-from-primary.bash
@@ -58,6 +58,11 @@ sudo -iu postgres psql -c "SELECT pg_switch_xlog();" > /dev/null 2>&1
 sudo -iu postgres psql -c "SELECT pg_switch_wal();" > /dev/null 2>&1
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=/var/lib/pgbackrest/archive --repo-host="backup-srv" --repo-host-user=postgres | cut -f1 -d"-" > $RESULTS_DIR/archives-ok.out
 
+# --service=archives --ignore-archived-before
+echo "--service=archives --ignore-archived-before"
+sudo -iu postgres psql -c "SELECT pg_sleep(2);" > /dev/null 2>&1
+$PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=/var/lib/pgbackrest/archive --repo-host="backup-srv" --repo-host-user=postgres --ignore-archived-before=1s > $RESULTS_DIR/archives-ignore-before.out
+
 # --service=archives --ignore-archived-after
 echo "--service=archives --ignore-archived-after"
 $PLUGIN_PATH/check_pgbackrest --stanza=my_stanza --service=archives --repo-path=/var/lib/pgbackrest/archive  --repo-host="backup-srv" --repo-host-user=postgres --ignore-archived-after=1h > $RESULTS_DIR/archives-ignore-after.out